[TASK-211] <<fixed>> rename CertificateBanner to CertificatePageBanner
