+++ itp-front/src/components/CertificatePageBanner.vue
